restructured cm_main tests
