+++ forum/tests/tests_reply.py
@@ -1,6 +1,6 @@
 from django.urls import reverse
 from django.core.exceptions import ValidationError
-from cm_main.tests import TestFollowersMixin
+from cm_main.tests.tests_followers import TestFollowersMixin
 from forum.tests.tests_post import ForumTestCase
 from ..models import Message
